Changed package to com.github.piotrga.asyncdynamo (closes #22)
