+++ src/main/scala/com.github.piotrga/asyncdynamo/TracingAmazonDynamoDB.scala
@@ -1,4 +1,4 @@
-package asyncdynamo
+package com.github.piotrga.asyncdynamo
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
 import com.amazonaws.services.dynamodbv2.model._
