+++ src/main/scala/com.github.piotrga/asyncdynamo/package.scala
@@ -14,12 +14,14 @@
  *    limitations under the License.
  */
 
-package asyncdynamo
+package com.github.piotrga.asyncdynamo
 
 import akka.actor.ActorRef
 import akka.util.Timeout
 import com.amazonaws.services.dynamodbv2.model.AttributeValue
-import asyncdynamo.nonblocking.ColumnCondition
+
+// This project
+import nonblocking.ColumnCondition
 
 package object blocking {
   def Read[T](id: String, range: Option[String] = None)(implicit dynamo: ActorRef, timeout: Timeout, dyn: DynamoObject[T]) = nonblocking.Read(id,range).blockingExecute
@@ -43,4 +45,39 @@ package object blocking {
   def IsTableActive[T]()(implicit dynamo: ActorRef, timeout: Timeout, dyn: DynamoObject[T]) = nonblocking.IsTableActive[T]().blockingExecute
 
   def CreateTable[T](readThroughput: Long =5, writeThrougput: Long = 5)(implicit dynamo: ActorRef, timeout: Timeout, dyn: DynamoObject[T]) = nonblocking.CreateTable[T](readThroughput, writeThrougput).blockingExecute
+}
+
+package object functional {
+  /**
+   * Useful for paging, ie:
+   * {{{
+  def execute(implicit cassandra: ActorRef, timeout: FiniteDuration) : Stream[(String,String)]= {
+    Streams.unfold(""){
+      (marker : String) =>
+        val res = nonblocking.GetSlice(columnFamilyId, rowId, marker, "", batchSize).executeOn(cassandra).get.toSeq
+        val elements = if (res.headOption.isDefined && marker == res.head._1)
+          res.tail
+        else res
+
+        val newMarker = elements.lastOption.map(_._1)
+        (newMarker, elements)
+
+    }.flatten
+  }
+  }}}
+   * @param initial initial state, ie. for paging it could be empty marker or number zero for first page.
+   * @param f transforms some state to the next state and result, ie. in batching for a given marker(state) it produces collection of elements(batch) and next marker(new state)
+   * @tparam STATE type of the state, ie. for paging it could be a String.
+   * @tparam ELEMENT type of the element, ie. for paging it could be a Seq[A]
+   * @return stream of elements, ie. for paging it will be Stream[ Seq[A] ], so if you flatten it you will get Stream[A].
+   *
+   */
+  def unfold[STATE, ELEMENT](initial: STATE)(f: STATE => (Option[STATE], ELEMENT)): Stream[ELEMENT] = {
+    (f(initial) match {
+      case (Some(s), batch) => Stream.cons(batch, unfold(s)(f))
+      case (None, list) => Stream(list)
+    })
+  }
+
+
 }
